server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["vikyw89 <vikyw89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
llama-index = "^0.9.11"
trulens-eval = "^0.18.2"
google-cloud-aiplatform = "^1.36.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
